+++ reco_analysis/reco_analysis/app/credentials.yaml
@@ -1,0 +1,36 @@
+credentials:
+  usernames:
+    "12305811":
+      email: kmorris@gmail.com
+      failed_login_attempts: 0 # Will be managed automatically
+      logged_in: False # Will be managed automatically
+      name: Kevin Morris
+      password: abc # Will be hashed automatically
+    "14185111":
+      email: mowens@gmail.com
+      failed_login_attempts: 0 # Will be managed automatically
+      logged_in: False # Will be managed automatically
+      name: Michael Owens
+      password: abc # Will be hashed automatically
+    "10339317":
+      email: rhenderson@gmail.com
+      failed_login_attempts: 0 # Will be managed automatically
+      logged_in: False # Will be managed automatically
+      name: Robert Henderson
+      password: abc # Will be hashed automatically
+    "14807966":
+      email: greynolds@gmail.com
+      failed_login_attempts: 0 # Will be managed automatically
+      logged_in: False # Will be managed automatically
+      name: Gregory Reynolds
+      password: abc # Will be hashed automatically
+    "13912736":
+      email: jcollins@gmail.com
+      failed_login_attempts: 0 # Will be managed automatically
+      logged_in: False # Will be managed automatically
+      name: Jennifer Collins
+      password: abc # Will be hashed automatically
+cookie:
+  expiry_days: 30
+  key: some_signature_key # Must be string
+  name: some_cookie_name
